// src/user/dto/create-Parking.dto.ts
export class CreateParkingDto {
  name: string;
  address: string;
  pinCode: number;
  locality: string;
  city: string;
  state: string;
  paymentType: string;
  parkingSlotType: string;
  startNo: number;
  endNo: number;
  collectionType: string;
  bankAccountDetails: number;
  userId: string;

}